backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.121.1",
    "google-genai>=1.49.0",
    "groq>=0.11.0",
    "gradio-client>=1.3.0",
    "langgraph>=1.0.2",
    "numpy>=1.26.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "replicate>=1.0.0",
    "requests>=2.31.0",
    "yt-dlp>=2025.10.22",
    "huggingface-hub>=1.1.2",
]
